RaceLoadBalancer 自定义探测委托异常击穿竞速调用兜底；XmlGeneral Decimal 经 Double 中转丢精度修正

+++ DH.NCore/Serialization/Xml/XmlGeneral.cs
@@ -200,7 +200,8 @@ public class XmlGeneral : XmlHandlerBase
                 value = v.ToDateTime();
                 return true;
             case TypeCode.Decimal:
-                value = (Decimal)v.ToDouble();
+                // 直接转 Decimal 保真，经 Double 中转会丢失精度（Decimal 高达 28-29 位有效数字）
+                value = v.ToDecimal();
                 return true;
             case TypeCode.Double:
                 value = v.ToDouble();
